+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Add support of multiple `--filter` options.
+- Add `--output-file` option to save serial output into file.
 
 ## [0.1.0] - 2021-04-03
 ### Added